mini-os: arm: memory management
Based on an initial patch by Karim Raslan.


---

Changes since v6:

- fixed printk format types
- physical_address_offset is now uint32_t

+++ extras/mini-os/include/arm/arch_mm.h
@@ -17,10 +17,10 @@ extern uint32_t physical_address_offset;	/* Add this to a virtual address to get
 #define to_phys(x)                 (((paddr_t)(x)+physical_address_offset) & 0xffffffff)
 #define to_virt(x)                 ((void *)(((x)-physical_address_offset) & 0xffffffff))
 
-#define PFN_UP(x)    (((x) + PAGE_SIZE-1) >> L1_PAGETABLE_SHIFT)
-#define PFN_DOWN(x)    ((x) >> L1_PAGETABLE_SHIFT)
+#define PFN_UP(x)    (unsigned long)(((x) + PAGE_SIZE-1) >> L1_PAGETABLE_SHIFT)
+#define PFN_DOWN(x)    (unsigned long)((x) >> L1_PAGETABLE_SHIFT)
 #define PFN_PHYS(x)    ((uint64_t)(x) << L1_PAGETABLE_SHIFT)
-#define PHYS_PFN(x)    ((x) >> L1_PAGETABLE_SHIFT)
+#define PHYS_PFN(x)    (unsigned long)((x) >> L1_PAGETABLE_SHIFT)
 
 #define virt_to_pfn(_virt)         (PFN_DOWN(to_phys(_virt)))
 #define virt_to_mfn(_virt)         (PFN_DOWN(to_phys(_virt)))
